Move pattern detection to its own module
