+++ src/format.rs
@@ -4,8 +4,8 @@ use crate::args::*;
 use crate::ignore::*;
 use crate::indent::*;
 use crate::logging::*;
+use crate::pattern::Pattern;
 use crate::read::*;
-use crate::regexes::{ENV_BEGIN, ENV_END, ITEM, RE_SPLITTING};
 use crate::subs::*;
 use crate::verbatim::*;
 use crate::wrap::*;
@@ -246,36 +246,6 @@ impl State {
     }
 }
 
-/// Record whether a line contains certain patterns to avoid recomputing
-#[derive(Default)]
-pub struct Pattern {
-    /// Whether a begin environment pattern is present
-    pub contains_env_begin: bool,
-    /// Whether an end environment pattern is present
-    pub contains_env_end: bool,
-    /// Whether an item pattern is present
-    pub contains_item: bool,
-    /// Whether a splitting pattern is present
-    pub contains_splitting: bool,
-}
-
-impl Pattern {
-    /// Check if a string contains patterns
-    pub fn new(s: &str) -> Self {
-        let mut pattern = Self::default();
-
-        // If splitting does not match, no patterns are present
-        if RE_SPLITTING.is_match(s) {
-            pattern.contains_env_begin = s.contains(ENV_BEGIN);
-            pattern.contains_env_end = s.contains(ENV_END);
-            pattern.contains_item = s.contains(ITEM);
-            pattern.contains_splitting = true;
-        }
-
-        pattern
-    }
-}
-
 /// Ensure that the indentation returns to zero at the end of the file
 const fn indents_return_to_zero(state: &State) -> bool {
     state.indent.actual == 0
@@ -302,19 +272,4 @@ pub fn run(args: &Args, logs: &mut Vec<Log>) -> u8 {
         }
     }
     exit_code
-}
-
-#[cfg(test)]
-mod tests {
-    use super::Pattern;
-
-    #[test]
-    fn new_pattern() {
-        let pattern =
-            Pattern::new("\\begin{enumerate} \\end{enumerate} \\item ");
-        assert!(pattern.contains_env_begin);
-        assert!(pattern.contains_env_end);
-        assert!(pattern.contains_item);
-        assert!(pattern.contains_splitting);
-    }
 }
