+++ src/verbatim.rs
@@ -2,6 +2,7 @@
 
 use crate::format::*;
 use crate::logging::*;
+use crate::pattern::Pattern;
 use crate::regexes::*;
 use log::Level::Warn;
